Feature/refactor admin data (#66)
* Refactor admin data
