+++ .dockerignore
@@ -24,5 +24,6 @@
 certs
 logs
 payloads
+data/postgres_data
 LICENSE
 README.md
